feat(ui): collapsible tab strip, chevron layering, active index bold
- Collapse sidebar to zero width; show chevrons only on hover; tight hit
  rects for expand/collapse so splitter drag and left-edge terminal input
  remain usable; draw < after splitter highlight.
- Bold active tab index via synthetic glyph overlay; fix tab_strip_collapsed
  declared before first layout_terms use.
- terminal_ui: float casts for DrawTextEx positions; comment on logical font_size.

Made-with: Cursor

+++ src/c/terminal_ui.c
@@ -534,11 +534,12 @@ void render_terminal(GhosttyRenderState render_state,
 
             int italic_offset = style.italic ? (font_size / 6) : 0;
 
-            DrawTextEx(font, text, (Vector2){x + italic_offset, y},
+            DrawTextEx(font, text, (Vector2){(float)(x + italic_offset), (float)y},
                        (float)font_size, 0, ray_fg);
 
             if (style.bold) {
-                DrawTextEx(font, text, (Vector2){x + italic_offset + 1, y},
+                DrawTextEx(font, text,
+                           (Vector2){(float)(x + italic_offset + 1), (float)y},
                            (float)font_size, 0, ray_fg);
             }
